Fix build

+++ src/gui/src/settings/options-window.cpp
@@ -1274,52 +1274,52 @@ void OptionsWindow::save()
 	//settings->setValue("Zoom/defaultButtonLayout", ui->checkZoomDefaultButtonLayout->isChecked());
 	QList<ButtonSettings> buttons;
 
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonPrev, ui->spinButtonPrev->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonPrev, (unsigned short) ui->spinButtonPrev->value(),
 		ui->lineButtonPrev->text(),
 		ui->checkButtonPrev->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonPrev->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonNext, ui->spinButtonNext->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonNext, (unsigned short) ui->spinButtonNext->value(),
 		ui->lineButtonNext->text(),
 		ui->checkButtonNext->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonNext->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonDetails, ui->spinButtonDetails->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonDetails, (unsigned short) ui->spinButtonDetails->value(),
 		ui->lineButtonDetails->text(),
 		ui->checkButtonDetails->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonDetails->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonSaveAs, ui->spinButtonSaveAs->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonSaveAs, (unsigned short) ui->spinButtonSaveAs->value(),
 		ui->lineButtonSaveAs->text(),
 		ui->checkButtonSaveAs->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonSaveAs->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonSave, ui->spinButtonSave->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonSave, (unsigned short) ui->spinButtonSave->value(),
 		ui->lineButtonSave->text(),
 		ui->checkButtonSave->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonSave->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonSaveNQuit, ui->spinButtonSaveNQuit->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonSaveNQuit, (unsigned short) ui->spinButtonSaveNQuit->value(),
 		ui->lineButtonSaveNQuit->text(),
 		ui->checkButtonSaveNQuit->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonSaveNQuit->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonOpen, ui->spinButtonOpen->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonOpen, (unsigned short) ui->spinButtonOpen->value(),
 		ui->lineButtonOpen->text(),
 		ui->checkButtonOpen->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonOpen->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonSave | Ui::IsFavoriteButton, ui->spinButtonSave->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonSave | Ui::IsFavoriteButton, (unsigned short) ui->spinButtonSave->value(),
 		ui->lineButtonSaveFav->text(),
 		ui->checkButtonSaveFav->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonSaveFav->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonSaveNQuit | Ui::IsFavoriteButton, ui->spinButtonSaveNQuit->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonSaveNQuit | Ui::IsFavoriteButton, (unsigned short) ui->spinButtonSaveNQuit->value(),
 		ui->lineButtonSaveNQuitFav->text(),
 		ui->checkButtonSaveNQuitFav->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonSaveNQuitFav->checkState() == Qt::PartiallyChecked ? true : false
 	});
-	buttons.append((ButtonSettings){ButtonSettings::Zoom, Ui::IsButtonOpen | Ui::IsFavoriteButton, ui->spinButtonOpen->value(),
+	buttons.append(ButtonSettings {ButtonSettings::Zoom, Ui::IsButtonOpen | Ui::IsFavoriteButton, (unsigned short) ui->spinButtonOpen->value(),
 		ui->lineButtonOpenFav->text(),
 		ui->checkButtonOpenFav->checkState() == Qt::Unchecked ? false : true,
 		ui->checkButtonOpenFav->checkState() == Qt::PartiallyChecked ? true : false
